backend/requirements.txt
flask
flask-sock
websocket-client
twilio
python-dotenv